Improved barogram performance.
Use selectors to generate the tick formatting functions.

The graph is no longer unnecessarily redrawn whenever the time index changes.
