+++ src/selectors/barogramSelectors.js
@@ -1,0 +1,72 @@
+import { createSelector } from 'reselect';
+import moment from 'moment';
+import { getTimestamps, getUtcOffset } from './timeSelectors';
+import { getAltitudes } from './altitudeSelectors';
+
+export const getBarogramData = createSelector(
+  getTimestamps,
+  getAltitudes,
+  (timestamps, altitudes) => {
+    let series = [];
+
+    timestamps.forEach((t, index) => {
+      series.push([t.unix(), altitudes.get(index)]);
+    });
+
+    return [series];
+  });
+
+export const getTickFormatter = createSelector(
+  getUtcOffset,
+  utcOffset => (t => moment.unix(t).utcOffset(utcOffset).format('HH:mm'))
+);
+
+function chooseTickInterval(durationMinutes) {
+  let interval;
+  if (durationMinutes <= 10) {
+    interval = 1;
+  }
+  if (durationMinutes <= 50) {
+    interval = 5;
+  }
+  else if (durationMinutes <= 100) {
+    interval = 10;
+  }
+  else if (durationMinutes <= 150) {
+    interval = 15;
+  }
+  else if (durationMinutes <= 300) {
+    interval = 30;
+  }
+  else if (durationMinutes <= 600) {
+    interval = 60;
+  }
+  else {
+    interval = 120;
+  }
+
+  return interval;
+}
+
+export const getTickGenerator = createSelector(
+  getUtcOffset,
+  utcOffset => axis => {
+    const ticks = [];
+    const startMoment = moment.unix(axis.min).utcOffset(utcOffset);
+    const endMoment = moment.unix(axis.max).utcOffset(utcOffset);
+    const durationMinutes = endMoment.diff(startMoment, 'minutes');
+    const interval = chooseTickInterval(durationMinutes);
+
+    const tick = startMoment.clone()
+      .minutes(0)
+      .seconds(0);
+
+    while (tick < endMoment) {
+      if (tick > startMoment) {
+        ticks.push(tick.unix());
+      }
+      tick.add(interval, 'minutes');
+    }
+
+    return ticks;
+  });
